@@ -82,7 +82,7 @@
     "@thorchain/asgardex-theme": "^0.1.1",
     "@thorchain/asgardex-util": "^0.10.0",
     "@thorchain/ledger-thorchain": "^0.1.0-alpha.2",
-    "@types/electron-devtools-installer": "^2.2.0",
+    "@types/electron-devtools-installer": "^2.2.2",
     "@xchainjs/xchain-binance": "^5.4.3",
     "@xchainjs/xchain-bitcoin": "^0.18.2",
     "@xchainjs/xchain-bitcoincash": "^0.13.1",
@@ -95,46 +95,45 @@
     "@xchainjs/xchain-terra": "^0.1.2",
     "@xchainjs/xchain-thorchain": "^0.24.1",
     "@xchainjs/xchain-util": "^0.7.1",
-    "antd": "^4.16.13",
+    "antd": "^4.20.2",
     "axios": "^0.25.0",
     "bchaddrjs": "^0.5.2",
     "bech32-buffer": "^0.2.0",
-    "bignumber.js": "^9.0.1",
+    "bignumber.js": "^9.0.2",
     "bitcoinjs-lib": "^5.2.0",
-    "caniuse-lite": "^1.0.30001336",
+    "caniuse-lite": "^1.0.30001338",
     "chart.js": "^3.7.1",
     "coininfo": "^5.1.0",
     "cosmos-client": "0.39.2",
     "dayjs": "^1.10.7",
     "electron-debug": "^3.2.0",
-    "electron-log": "^4.4.6",
+    "electron-log": "^4.4.7",
     "electron-updater": "^5.0.1",
     "electron-window-state": "^5.0.3",
     "ethers": "^5.4.7",
     "fp-ts": "^2.11.5",
     "fp-ts-rxjs": "^0.6.15",
-    "fs-extra": "^10.0.0",
+    "fs-extra": "^10.1.0",
     "io-ts": "^2.2.16",
     "io-ts-types": "^0.5.16",
     "lodash.shuffle": "^4.2.0",
-    "observable-hooks": "^4.1.0",
-    "qrcode": "^1.4.4",
-    "react": "^17.0.2",
+    "observable-hooks": "^4.2.0",
+    "qrcode": "^1.5.0",
+    "react": "^18.1.0",
     "react-chartjs-2": "^4.1.0",
     "react-circular-progressbar": "^2.0.4",
-    "react-dom": "^17.0.2",
-    "react-intl": "^5.20.13",
+    "react-dom": "^18.1.0",
+    "react-intl": "^5.25.1",
     "react-router-dom": "^6.3.0",
     "react-scripts": "^4.0.3",
-    "react-test-renderer": "^17.0.2",
+    "react-test-renderer": "^18.1.0",
     "rxjs": "^6.6.7",
-    "styled-components": "^5.3.1",
+    "styled-components": "^5.3.5",
     "styled-theme": "^0.3.3",
     "uuid": "^8.3.2",
     "wif": "^2.0.6"
   },
   "devDependencies": {
-    "@hot-loader/react-dom": "^17.0.1",
     "@react-theming/storybook-addon": "^1.1.1",
     "@storybook/addon-actions": "^6.3.12",
     "@storybook/addon-backgrounds": "^6.3.12",
@@ -151,27 +150,27 @@
     "@types/bchaddrjs": "^0.4.0",
     "@types/fs-extra": "^9.0.13",
     "@types/jest": "^27.0.2",
-    "@types/lodash.shuffle": "^4.2.6",
+    "@types/lodash.shuffle": "^4.2.7",
     "@types/node": "^16.11.7",
-    "@types/qrcode": "^1.4.1",
-    "@types/react": "^17.0.30",
-    "@types/react-dom": "^17.0.9",
+    "@types/qrcode": "^1.4.2",
+    "@types/react": "^18.0.8",
+    "@types/react-dom": "^18.0.3",
     "@types/react-intl": "^3.0.0",
-    "@types/styled-components": "^5.1.15",
+    "@types/styled-components": "^5.1.25",
     "chalk": "^5.0.1",
     "concurrently": "^7.1.0",
     "cross-env": "^7.0.3",
     "electron": "^18.1.0",
     "electron-builder": "^23.0.3",
-    "electron-devtools-installer": "^2.2.2",
+    "electron-devtools-installer": "^3.2.0",
     "electron-notarize": "^1.2.1",
     "eslint-config-prettier": "^8.5.0",
     "eslint-plugin-prettier": "^4.0.0",
     "git-revision-webpack-plugin": "^5.0.0",
     "husky": "^7.0.4",
-    "lint-staged": "^12.4.0",
+    "lint-staged": "^12.4.1",
     "node-abi": "^3.15.0",
-    "nodemon": "^2.0.15",
+    "nodemon": "^2.0.16",
     "prettier": "^2.6.2",
     "rimraf": "^3.0.2",
     "source-map-explorer": "^2.5.2",
@@ -181,7 +180,7 @@
     "ts-jest": "^27.1.4",
     "ts-loader": "^8.3.0",
     "ts-node": "^10.7.0",
-    "typescript": "^4.6.3",
+    "typescript": "^4.6.4",
     "wait-on": "^6.0.1",
     "webpack-cli": "^4.9.2",
     "webpack-node-externals": "^3.0.0"
