@@ -1,12 +1,17 @@
 import React from 'react'
 
-import ReactDOM from 'react-dom'
-
 import './index.css'
+import { createRoot } from 'react-dom/client'
+
 import { App } from './App'
 import * as serviceWorker from './serviceWorker'
 
-ReactDOM.render(<App />, document.getElementById('root'))
+// React 18 introduces a new root API
+// @see https://reactjs.org/blog/2022/03/08/react-18-upgrade-guide.html#updates-to-client-rendering-apis
+const container = document.getElementById('root')
+// eslint-disable-next-line @typescript-eslint/no-non-null-assertion
+const root = createRoot(container!) // createRoot(container!) if you use TypeScript
+root.render(<App />)
 
 // If you want your app to work offline and load faster, you can change
 // unregister() to register() below. Note this comes with some pitfalls.
