@@ -36,12 +36,12 @@ export const FilterMenu = <T extends unknown>(props: Props<T>): JSX.Element => {
   const [searchTerm, setSearchTerm] = useState('')
 
   const handleClick: MenuProps['onClick'] = useCallback(
-    (event) => {
+    ({ key }: { key: string }) => {
       // (Rudi) bail if this is triggered by the search menu item
-      if (!event || !event.key || event.key === '_search') return
+      if (!key || key === '_search') return
 
       setSearchTerm('')
-      onSelect(event.key)
+      onSelect(key)
     },
     [onSelect]
   )
