@@ -29,7 +29,7 @@ const initialContext: AppContextValue = {
 
 const AppContext = createContext<AppContextValue | null>(null)
 
-export const AppProvider: React.FC = ({ children }): JSX.Element => (
+export const AppProvider: React.FC<{ children: React.ReactNode }> = ({ children }): JSX.Element => (
   <AppContext.Provider value={initialContext}>{children}</AppContext.Provider>
 )
 
