@@ -63,7 +63,7 @@ const getAntdLocale = (locale: Locale) => {
   }
 }
 
-export const I18nProvider: React.FC = ({ children }): JSX.Element => {
+export const I18nProvider: React.FC<{ children: React.ReactNode }> = ({ children }): JSX.Element => {
   const locale = useObservableState(locale$, initialLocale)
   const messages = useMemo(() => getMessagesByLocale(locale), [locale])
   const antdLocale = getAntdLocale(locale)
