@@ -61,7 +61,7 @@ const initialContext: WalletContextValue = {
 }
 const WalletContext = createContext<Option<WalletContextValue>>(none)
 
-export const WalletProvider: React.FC = ({ children }): JSX.Element => (
+export const WalletProvider: React.FC<{ children: React.ReactNode }> = ({ children }): JSX.Element => (
   <WalletContext.Provider value={some(initialContext)}>{children}</WalletContext.Provider>
 )
 
